axum

+++ hello-axum/migrations/20250106054733_user.sql
@@ -1,4 +1,5 @@
 create table if not exists user(
-    id         INTEGER PRIMARY KEY NOT NULL,
+    id         BIGINT PRIMARY KEY NOT NULL,
     name       TEXT                NOT NULL
 )
+
